update code flow 13 04 23

+++ paymentService/src/main/java/com/travel/paymentService/entity/PresentmentRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class PresentmentRequest {
-    private Integer packageId;
-    private Integer accountId;
-    private Integer count;
+    private Long packageId;
+    private Long accountId;
+    private Long count;
 }
