+++ paymentService/src/main/java/com/travel/paymentService/entity/PresentmentResponse.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class PresentmentResponse {
-    private Integer paymentHistoryId;
+    private Long paymentHistoryId;
     private String name;
     private String account;
     private String packageName;
