+++ paymentService/src/main/java/com/travel/paymentService/model/TravelPackage.java
@@ -10,9 +10,9 @@ public class TravelPackage {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @Column(name = "name")
     private String name;
     @Column(name = "price")
-    private Integer price;
+    private Long price;
 }
