+++ paymentService/src/main/java/com/travel/paymentService/repository/PaymentHistoryDBRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PaymentHistoryDBRepository extends JpaRepository<PaymentHistory,Integer> {
+public interface PaymentHistoryDBRepository extends JpaRepository<PaymentHistory,Long> {
 }
