+++ paymentService/src/main/java/com/travel/paymentService/repository/TravelPackageDBRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TravelPackageDBRepository extends JpaRepository<TravelPackage,Integer> {
+public interface TravelPackageDBRepository extends JpaRepository<TravelPackage,Long> {
 }
